pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["g_workspace_mcp"]

[project]
name = "g-workspace-mcp"
version = "0.1.0"
description = "Google Workspace MCP Server for Claude Code, Cursor, and Gemini CLI"
readme = "README.md"
keywords = ["mcp", "google-workspace", "drive", "gmail", "calendar", "sheets"]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "structlog>=24.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[project.scripts]
g-workspace-mcp = "g_workspace_mcp.src.cli:main"

[project.urls]
Repository = "https://github.com/jbendavi/g-workspace-mcp"

[tool.ruff]
target-version = "py311"
line-length = 100
lint.select = ["E", "W", "F", "I"]
lint.ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
